# perf: cache json initialization