I was doing some perf for `npm run` and I found the `sortTranspositions` was taking 5ms to run.

![image](https://github.com/jslicense/spdx-correct.js/assets/12551007/4f0c601d-7789-468f-bff4-ff1e95ac63db)

So, I changed the code to be like `spdx-license-ids` and generated a JSON with the calculated `transpositions` and also with the `last-resorts`.

I also lazy load the dependencies to only import when needed, so we do not introduce any perf penalty for those who import this module but will not use it immediately.

In this way, we remove one dependency and we also improve the initialization of this module:

Before:

```bash
$ hyperfine --warmup 3 "node ./index.js"                                                                                                                                                            [22:38:45]
Benchmark 1: node ./index.js
  Time (mean ± σ):     109.3 ms ±   1.1 ms    [User: 95.4 ms, System: 38.1 ms]
  Range (min … max):   107.2 ms … 111.9 ms    27 runs
```


After:

```bash
$ hyperfine --warmup 3 "node ./index.js"                                                                                                                                  [22:36:37]
Benchmark 1: node ./index.js
  Time (mean ± σ):     101.0 ms ±   1.1 ms    [User: 88.5 ms, System: 34.1 ms]
  Range (min … max):    99.2 ms … 103.6 ms    29 runs
```