@kemitchell 2024-04-06
Are there performance concerns reported upstream? Or is this optimization for its own sake?

Removing dependency on spdx-license-ids may work a small load-time improvement, but it will also prevent relatively frequent updates to that package from flowing through to this package via `npm upgrade`, without lockstep releases of `spdx-correct`. In npm in particular, I expect there are also other edges to spdx-license-ids in the dependency tree, so essentially vendoring it under spdx-correct and continuing to depend on it elsewhere will lead to duplication in installs and distributed build archives.

I may have fallen behind on changes to the various `npm` subcommands. My impression was that npm continues to use this package for `npm init`, to help users choose valid `package.json` `license` property values. If the load time is a concern, why is `npm run` loading this package at all?