@H4ad 2024-04-06
> Are there performance concerns reported upstream? Or is this optimization for its own sake?

There is no concern upstream, I'm just a dev with free time and willing to speedup npm run.

> but it will also prevent relatively frequent updates to that package from flowing through to this package via

That's a valid concern, but I think you can solve this by using `dependabot` to be always updated with the latest version, to save some `ms` from the initialization, I think it is worth it.

> In npm in particular, I expect there are also other edges to spdx-license-ids

You are right on this one:

```
npm@10.5.1 /home/h4ad/Projects/opensource/performance-analysis/npm-cli
├─┬ init-package-json@6.0.2
│ └─┬ validate-npm-package-license@3.0.4
│   └─┬ spdx-correct@3.2.0
│     └── spdx-license-ids@3.0.17 deduped
└─┬ spdx-expression-parse@3.0.1
  └── spdx-license-ids@3.0.17
```

> If the load time is a concern, why is npm run loading this package at all?

Good question, from what I saw, is just for legacy purposes, so I will try to lazy load the initialization of this module.

But currently, importing this module will have a penalty of `5ms` without even calling the exported function.

---

In case you don't want to ship the JSONs, I will totally understand! If so, can we at least lazy load the initialization to when this module is being called? I think this will be the ideal world, to leave the penalty only when this module is actually being used.