@kemitchell 2024-04-06
If you're hunting perf gains, I'm guessing the real prize is in npm. I don't know what legacy `npm run` preserves by loading this package, but I maintain this package, and I have no idea what `npm run` has to do with it!

As for this package alone, I'm disinclined to make a chore of lockstep releases, especially premature of any perf complaints. It's not a pure win, given how often this package gets bundled with other spdx-related packages. I'm not if this is being used in any Web client-side apps, but space matters, too.

Frankly, mostly it's my time that I value. I try to avoid making chores for myself.