@H4ad 2024-04-06
That's okay, totally valid reasons.

Before closing this PR, do you think is valid to create a PR to lazy load the initialization of `transpositions` and `lastResorts`?