@kemitchell 2024-04-06
I'm not exactly sure I know what you mean. But if you're feeling inspired to PR while you've got it fresh in mind, I'll earnestly review your proposal. Maybe not this evening—it's 1930 here—but when I can.