@H4ad 2024-04-06
I created the PR at https://github.com/jslicense/spdx-correct.js/pull/44

Take your time, I really appreciate your review, as a perf guy, I just look at the perf gains without thinking much about the other "costs" of the code.

And for me it's almost midnight, so I'm going to rest now ahahha